General cleanup, remove old constructs

+++ lib/Crypto/PublicKey/ECC.py
@@ -48,7 +48,7 @@ class _Curve(object):
     pass
 
 _curve = _Curve()
-_curve.p = Integer(0xffffffff00000001000000000000000000000000ffffffffffffffffffffffffL)
+_curve.p = Integer(0xffffffff00000001000000000000000000000000ffffffffffffffffffffffff)
 _curve.b = Integer(0x5ac635d8aa3a93e7b3ebbd55769886bc651d06b0cc53b0f63bce3c3e27d2604b)
 _curve.order = Integer(0xffffffff00000000ffffffffffffffffbce6faada7179e84f3b9cac2fc632551)
 _curve.Gx = Integer(0x6b17d1f2e12c4247f8bce6e563a440f277037d812deb33a0f4a13945d898c296)
